feat: Use detected quadrilateral in preview
In preview mode, i.e. scanning a document with the camera, the
quadrilateral was first detected, and then computed again from the
taken picture. This takes useless resources and was sometimes causing a
missing detection, supposively because another camera frame was used.
We now save the detected quadrilateral in preview and use it afterwards.